requirements.txt
Django==5.0.4
django-countries==7.6.1
django-crispy-forms==2.1
django-extensions==3.2.3
django-filter==24.2
django-js-asset==2.2.0
django-leaflet==0.29.1
django-sass-processor==1.4
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
django-rest-swagger==2.2.0
drf-yasg==1.21.7
folium==0.16.0
geographiclib==2.0
geopy==2.4.1
gunicorn==22.0.0
h11==0.14.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
psycopg==3.1.18
psycopg2-binary==2.9.9
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.0
requests-oauthlib==2.0.0
shapely==2.0.4
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
whitenoise==6.6.0
xyzservices==2024.4.0